Crosstab reporting solution
Sample code that solves the problem of tranposing row-wise report data into a fixed width report design. Theoretically there can be an unlimited number of columns in the transposition, and in this case the intent is to slot them into a fixed six column layout that can be printed in portrait mode. Any columns/pages that extend beyond the fixed layout are tracked using a horizontal page attribute.  So for example if the report required a total of 6 pages (picture a matrix of 3 horizontal pages and 2 vertical pages), it would ultimately appear in the UI as only a single page wide, using a sequence of 1-1, 1-2, 1-3, 2-1, 2-2, 2-3 (note each horizontal page appears before moving on to the next vertical page). The printed output would match the UI.

Note the code is not compilable - there are many missing classes and imports - it is simply intended for demonstration purposes.